Function algorithm::divide_and_conquer::max_subarray

pub fn max_subarray<'a, T>(arr: &'a Vec<T>) -> &'a [T] where
    T: Add + AddAssign + Clone + Ord, 

Find maxinum nonempty subarray whose sum is largest.

Let’s think about how we might solve the maximum-subarray problem using the divide-and-conquer technique. Suppose we want to find a maximum subarray of the subarray A[low..high]. Divide-and-conquer suggests that we divide the subarray into two subarrays of as equal size as possible. That is, we find the midpoint, say mid, of the subarray, and consider the subarrays A[low..mid] and A[mid+1..high]